+++ app/javascript/components/record-form/form/subforms/subform-header/components/list-item-text-secondary/component.jsx
@@ -17,12 +17,16 @@ const Component = ({ associatedViolations, violationsIDs, renderSecondaryText = 
   const violationsKeys = buildAssociatedViolationsKeys(associatedViolations, violationsIDs);
 
   return (
-    <div id="list-item-text-secondary-violations">
+    <div id="list-item-text-secondary-violations" data-testid="list-item-text-secondary-violations">
       <h4 className={css.title}>{i18n.t("incident.violation.associated_violations")}</h4>
       {violationsKeys.map(violationKey => {
         return (
           <div className={css.item}>
-            <div id={`id-div-list-item-seconday-${violationKey}`} key={`key-list-item-seconday-${violationKey}`}>
+            <div
+              data-testid="volation-keys"
+              id={`id-div-list-item-seconday-${violationKey}`}
+              key={`key-list-item-seconday-${violationKey}`}
+            >
               {i18n.t(`incident.violation.types.${violationKey}`)}
             </div>
           </div>
